Backend/requirements.txt
Flask==3.1.3
python-dotenv>=1.0.0
flask-cors==4.0.1
ultralytics>=8.3.0
opencv-python>=4.9.0
Pillow>=10.3.0
numpy>=1.26.4
python-magic>=0.4.27
